refactor: Remove LocationCard.vue and LocaltionView.vue components
